+++ src/main/java/com/excilys/service/ComputerDtoService.java
@@ -17,8 +17,7 @@ public interface ComputerDtoService {
    * Method that check the information in the computerDTO and then call the updateComputer method
    * from ComputerService.
    *
-   * @param computerDto
-   *          the computer dto
+   * @param computerDto the computer dto
    * @see ComputerDTO, ComputerService
    */
   public void updateComputer(ComputerDto computerDto);
@@ -27,37 +26,33 @@ public interface ComputerDtoService {
    * Method that check the name and dates given in parameters and then call the insertComputer
    * method from the ComputerService.
    *
-   * @param computerDto
-   *          the computer dto
+   * @param computerDto the computer dto
    * @return The id of the computer we just inserted.
    * @see ComputerDTO, ComputerService
    */
-  public int insertComputer(ComputerDto computerDto);
+  public long insertComputer(ComputerDto computerDto);
 
   /**
    * Method that calls the deleteComputer method from the ComputerService.
    * 
-   * @param id
-   *          The id of the computer we want to delete.
+   * @param id The id of the computer we want to delete.
    * @see ComputerDTO, ComputerService
    */
-  public void deleteComputer(int id);
+  public void deleteComputer(long id);
 
   /**
    * Method that calls the getById method from the ComputerService.
    * 
-   * @param id
-   *          The id of the computer we want to retrieve
+   * @param id The id of the computer we want to retrieve
    * @return The computer DTO that matches the given id or null
    * @see ComputerDTO, ComputerService
    */
-  public ComputerDto getById(int id);
+  public ComputerDto getById(long id);
 
   /**
    * Method that calls the getByName method from the ComputerService.
    * 
-   * @param name
-   *          The name of the computer(s) we want to retrieve.
+   * @param name The name of the computer(s) we want to retrieve.
    * @return The list of computers DTO that matches the given name (or an empty list).
    * @see ComputerDTO, ComputerService
    */
@@ -84,9 +79,8 @@ public interface ComputerDtoService {
    * Method that calls the getNbComputer from the ComputerService.
    * 
    * @param queryParameters
-   * 
    * @return The number of computer in the database.
    * @see ComputerService
    */
-  public int getCount(QueryParameters queryParameters);
+  public long getCount(QueryParameters queryParameters);
 }
